Move import mapping handling to a concern.
