+++ app/models/ontology.rb
@@ -14,6 +14,7 @@ class Ontology < ActiveRecord::Base
   include Ontology::FileExtensions
   include Ontology::HetsOptions
   include Ontology::Import
+  include Ontology::ImportMappings
   include Ontology::Mappings
   include Ontology::Oops
   include Ontology::OwlClasses
@@ -81,18 +82,6 @@ class Ontology < ActiveRecord::Base
     name? ? iri : nil
   end
 
-  def is_imported?
-    import_mappings.present?
-  end
-
-  def is_imported_from_other_repository?
-    import_mappings_from_other_repositories.present?
-  end
-
-  def imported_by
-    import_mappings.map(&:source)
-  end
-
   def destroy_with_parent(user)
     if parent
       repository.delete_file(parent.path, user,
@@ -128,19 +117,9 @@ class Ontology < ActiveRecord::Base
     !is_imported_from_other_repository?
   end
 
-  def imported_ontologies
-    fetch_mappings_by_kind(self, 'import')
-  end
-
   def contains_logic_translations?
     query, args = contains_logic_translations_query(self)
     pluck_select([query, *args], :logically_translated).size > 1
-  end
-
-  def direct_imported_ontologies
-    ontology_ids = Mapping.where(target_id: self, kind: 'import').
-      pluck(:source_id)
-    Ontology.where(id: ontology_ids)
   end
 
   def combined_sentences
@@ -211,15 +190,5 @@ class Ontology < ActiveRecord::Base
     else
       repository.path_exists?(path, commit_oid)
     end
-  end
-
-  protected
-
-  def import_mappings
-    Mapping.where(source_id: id, kind: 'import')
-  end
-
-  def import_mappings_from_other_repositories
-    import_mappings.select { |l| l.target.repository != repository }
   end
 end
